pyproject.toml
[tool.poetry]
name = "simplecel"
version = "0.2.10"
description = "Offline Excel-like app with no formula conversion, but with image/markdown/HTML support."
authors = ["Pacharapol Withayasakpunt <patarapolw@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/patarapolw/simplecel"
homepage = "https://github.com/patarapolw/simplecel"

keywords = ['excel', 'handsontable', 'pyexcel']

[tool.poetry.dependencies]
python = ">=3"
pyexcel = "^0.5.8"
flask = "^1.0"
click = "^6.7"
"ruamel.yaml" = "^0.15.48"

[tool.poetry.dev-dependencies]
pytest = "^3.7"
python-dotenv = "^0.9.0"
#pyexcel-xlsx = "^0.5.6"
#pyexcel-export = "^0.3.5"
pyexcel-xls = "^0.5.7"
pyexcel-xlsxwx = "^0.1.6"
#pyexcel-xlsxwx = {path = "../pyexcel-xlsxwx"}

[tool.poetry.scripts]
simplecel = 'simplecel.__main__:cli'
